some fixes for deadline model
